# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Canvica Cleaning Services CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs with optional filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, in_progress, completed, cancelled, late]
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: customerId
          in: query
          schema:
            type: number
        - name: employeeId
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Jobs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
  /jobs/{id}/status:
    patch:
      operationId: updateJobStatus
      tags: [jobs]
      summary: Update job status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobStatusUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"

  # ─── QUOTES ─────────────────────────────────────────────────────────────────
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, sent, paid, overdue, cancelled]
        - name: customerId
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Invoices list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /invoices/{id}/send:
    post:
      operationId: sendInvoice
      tags: [invoices]
      summary: Send invoice to customer by email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Invoice sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}/pay:
    patch:
      operationId: markInvoicePaid
      tags: [invoices]
      summary: Mark invoice as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRecord"
      responses:
        "200":
          description: Paid invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  # ─── EMPLOYEES ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, customerId, customerName, serviceType, cleaningType, status, scheduledDate, scheduledTime, price, createdAt]
      properties:
        id:
          type: number
        customerId:
          type: number
        customerName:
          type: string
        customerAddress:
          type: ["string", "null"]
        serviceType:
          type: string
          enum: [residential, commercial]
        cleaningType:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled, late]
        scheduledDate:
          type: string
          format: date
        scheduledTime:
          type: string
        durationHours:
          type: ["number", "null"]
        price:
          type: number
        assignedEmployeeIds:
          type: array
          items:
            type: number
        assignedEmployeeNames:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        completionNotes:
          type: ["string", "null"]
        customerRating:
          type: ["number", "null"]
        isRecurring:
          type: boolean
        recurrencePattern:
          type: ["string", "null"]
        invoiceId:
          type: ["integer", "null"]
        quoteId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    JobInput:
      type: object
      required: [customerId, serviceType, cleaningType, scheduledDate, scheduledTime, price]
      properties:
        customerId:
          type: number
        serviceType:
          type: string
          enum: [residential, commercial]
        cleaningType:
          type: string
        scheduledDate:
          type: string
          format: date
        scheduledTime:
          type: string
        durationHours:
          type: number
        price:
          type: number
        assignedEmployeeIds:
          type: array
          items:
            type: number
        notes:
          type: string
        isRecurring:
          type: boolean
        recurrencePattern:
          type: string
        quoteId:
          type: number
    JobUpdate:
      type: object
      properties:
        serviceType:
          type: string
          enum: [residential, commercial]
        cleaningType:
          type: string
        scheduledDate:
          type: string
          format: date
        scheduledTime:
          type: string
        durationHours:
          type: number
        price:
          type: number
        assignedEmployeeIds:
          type: array
          items:
            type: number
        notes:
          type: string
        completionNotes:
          type: string
        customerRating:
          type: number
        isRecurring:
          type: boolean
        recurrencePattern:
          type: string
    JobStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled, late]
        completionNotes:
          type: string
        customerRating:
          type: number

    # ─── QUOTE SCHEMAS ────────────────────────────────────────────────────────
...
    QuoteLineItem:
      type: object
      required: [description, quantity, unitPrice, total]
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
...
    Invoice:
      type: object
      required: [id, customerId, customerName, status, lineItems, subtotal, tax, total, dueDate, createdAt]
      properties:
        id:
          type: number
        invoiceNumber:
          type: string
        customerId:
          type: number
        customerName:
          type: string
        customerEmail:
          type: ["string", "null"]
        jobId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/QuoteLineItem"
        subtotal:
          type: number
        taxRate:
          type: number
        tax:
          type: number
        total:
          type: number
        paidAmount:
          type: number
        paymentMethod:
          type: ["string", "null"]
        paymentDate:
          type: ["string", "null"]
          format: date
        dueDate:
          type: string
          format: date
        sentAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    InvoiceInput:
      type: object
      required: [customerId, lineItems, dueDate]
      properties:
        customerId:
          type: number
        jobId:
          type: number
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/QuoteLineItem"
        taxRate:
          type: number
        dueDate:
          type: string
          format: date
        notes:
          type: string
    InvoiceUpdate:
      type: object
      properties:
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/QuoteLineItem"
        taxRate:
          type: number
        dueDate:
          type: string
          format: date
        notes:
          type: string
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
    PaymentRecord:
      type: object
      required: [paidAmount, paymentMethod]
      properties:
        paidAmount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, card, bank_transfer, cheque, other]
        paymentDate:
          type: string
          format: date
        notes:
          type: string

    # ─── EMPLOYEE SCHEMAS ─────────────────────────────────────────────────────
...